Fix false positive for duplicate command definition inspection in if/else
